fix up build breaks
